Move find_version_string() helper fn to common module
